add PATO:0080001 quality of a fluid
There are a number of qualities that are shared between liquids and
gasses related to how they flow (e.g. viscosity). This new class is
created to capture those qualities. However, it is not used as a
superclass of quality of a liquid and quality of a gas because there
are qualities of a liquid such as vapor pressure that are not also
qualities of a gas (without some bending of meaning of vapor
pressure). Thus all qualities of a liquid are not qualities of a fluid
so quality of a fluid cannot be the superclass for liquid/gas
qualities.